Refactored server file to multiple files since its going to grow
